This is the first iteration of the model.
We find the "optimal" polynomial relationship for each predictor with our
dependent variable by repeatedly splitting our data set into training and test
sets, then iterating up through higher degree polynomials until the absolute
error in the training set increases. We do this 100 times in an attempt to
remove the effect of the random sampling of the train/test split.

We then create a model by adding together all the polynomials with degree of at
least 1. This model sucks. The error is at best 10 times the error of of a line
going through the average y value, and at worst nearly 100 times the error.
This massive error is likely due to extreme saturation -- by graphing our
predictions against our target, it is apparent we are not "connecting the dots"
for the vast majority of the data set (though perhaps we are for some).

Saturation should be reduced by introducing terms that take into account
relationships between predictors. It may also help if we reduce the number of
terms used in the model, so we could potentially build a model by hand-
selecting the "best" predictors (those with the least error?) and using those.
It may also be interesting just to see at what threshold the model starts to
suck.